+++ mayan/apps/templating/urls.py
@@ -1,6 +1,8 @@
 from django.urls import re_path
 
-from .api_views import APITemplateDetailView, APITemplateListView
+from .api_views import (
+    APIObjectTemplateSandboxActionView, APITemplateDetailView, APITemplateListView
+)
 
 from .views import ObjectTemplateSandboxView
 
@@ -13,6 +15,11 @@ urlpatterns = [
 ]
 
 api_urls = [
+    re_path(
+        route=r'^objects/(?P<app_label>[-\w]+)/(?P<model_name>[-\w]+)/(?P<object_id>\d+)/sandbox/$',
+        view=APIObjectTemplateSandboxActionView.as_view(),
+        name='object-template-sandbox'
+    ),
     re_path(
         route=r'^templates/$', view=APITemplateListView.as_view(),
         name='template-list'